+++ projects/iad-interface-admin/partner/src/public_api.ts
@@ -35,7 +35,6 @@ export * from './lib/data-table/models/actual-selection.model';
 export * from './lib/data-table/services/data-table-information.service';
 export * from './lib/data-table/services/settings-provider';
 export * from './lib/data-table/data-table/data-table-columns.service';
-export * from './lib/data-table/services/actual-selection-chain.service';
 export * from './lib/data-table/services/selection-buffer.service';
 export * from './lib/data-table/data-table/data-table.component';
 export * from './lib/data-table/services/actual-selection-chain.service';